Progressive Web App Update & Work Schedule Pause Option
This update includes the option for users to install Satis Tracker to their phone or computer for easier access. Users now have an option to pause their work schedule in case they are taking a vacation, using sick days, or for any other reason they will not be at work. If they use this option, their reporting streak will not reset but they will not be able to post new mood reports as long as their work schedule is paused.